refactor: flatten filters structure
