+++ lib/avo/filters/boolean_filter.rb
@@ -1,8 +1,6 @@
-require_relative '../filter'
-
 module Avo
   module Filters
-    class BooleanFilter < Avo::Filter
+    class BooleanFilter < BaseFilter
       self.template = 'avo/base/boolean_filter'
     end
   end
