--- spec/dummy/app/avo/filters/boolean_filter.rb
@@ -1,3 +1,0 @@
-class BooleanFilter < Avo::Filter
-  self.template = 'avo/base/boolean_filter'
-end
