--- spec/dummy/app/avo/filters/select_filter.rb
@@ -1,3 +1,0 @@
-class SelectFilter < Avo::Filter
-  self.template = 'avo/base/select_filter'
-end
